Time: 1 ms (98.54%), Space: 42.7 MB (42.88%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -3,7 +3,7 @@ class Solution {
         int low=0,high=x, ans=0;
         while(low<=high){
             int mid=low+(high-low)/2;
-            if(mid*mid==x)
+            if((long)mid*mid==x)
                 return mid;
             else if((long)mid*mid<x){
                 ans=mid;
